Adding support for X-CSRF-Token header
Other major web frameworks support the X-CSRF-Token header as a
way to validate the CSRF token. An example is the Sencha Connect
framework which powers most node.js applications. The url for
reference is http://www.senchalabs.org/connect/csrf.html The name
for the header is configurable with "csrf.header.name"